requirements.txt
amqp==2.6.0
asgiref==3.2.10
attrs==20.1.0
billiard==3.6.3.0
boto3==1.14.12
botocore==1.17.12
celery==4.4.6
cffi==1.14.0
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.4.0
django-rest-knox==4.1.0
django-ses==1.0.1
djangorestframework==3.11.0
docutils==0.15.2
future==0.18.2
gunicorn==20.0.4
iniconfig==1.0.1
jmespath==0.10.0
kombu==4.6.11
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.9.0
python-dateutil==2.8.1
python-http-client==3.2.7
pytz==2020.1
redis==3.5.3
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
toml==0.10.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.1.0
